+++ dolby_vision/src/rpu/extension_metadata/blocks/level255.rs
@@ -11,6 +11,7 @@ use super::{ExtMetadataBlock, ExtMetadataBlockInfo};
 #[repr(C)]
 #[derive(Debug, Default, Clone)]
 #[cfg_attr(feature = "serde_feature", derive(Deserialize, Serialize))]
+#[cfg_attr(feature = "serde_feature", serde(default))]
 pub struct ExtMetadataBlockLevel255 {
     pub dm_run_mode: u8,
     pub dm_run_version: u8,