+++ dolby_vision/src/rpu/vdr_dm_data.rs
@@ -388,9 +388,7 @@ impl VdrDmData {
                 }
             }
             ExtMetadataBlock::Level11(_) => self.replace_metadata_level(block),
-            ExtMetadataBlock::Level254(_) => {
-                bail!("Cannot replace automatically generated Level254 block")
-            }
+            ExtMetadataBlock::Level254(_) => self.replace_metadata_level(block),
             ExtMetadataBlock::Level255(_) => self.replace_metadata_level(block),
             ExtMetadataBlock::Reserved(_) => bail!("Cannot replace specific reserved block"),
         }